+++ src/idunno.Bluesky/Notifications/Agent/ListNotifications.cs
@@ -1,6 +1,8 @@
 // Copyright (c) Barry Dorrans. All rights reserved.
 // Licensed under the MIT License.
 
+using System.Diagnostics.CodeAnalysis;
+
 using idunno.AtProto;
 using idunno.Bluesky.Notifications;
 
@@ -13,7 +15,6 @@ public partial class BlueskyAgent
     /// </summary>
     /// <param name="limit">The maximum number of notifications to return. If specified this should be &gt;=1 and &lt;= 100.</param>
     /// <param name="cursor">An optional cursor. See https://atproto.com/specs/xrpc#cursors-and-pagination.</param>
-    /// <param name="seenAt">The date and time notifications were last checked.</param>
     /// <param name="reasons">An optional collection of <see cref="NotificationReason"/>s to limit the results to.</param>
     /// <param name="subscribedLabelers">A optional list of labeler <see cref="Did"/>s to accept labels from.</param>
     /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
@@ -24,7 +25,6 @@ public partial class BlueskyAgent
     public async Task<AtProtoHttpResult<NotificationCollection>> ListNotifications(
         int? limit = null,
         string? cursor = null,
-        DateTimeOffset? seenAt = null,
         IEnumerable<NotificationReason>? reasons = null,
         IEnumerable<Did>? subscribedLabelers = null,
         CancellationToken cancellationToken = default)
@@ -37,7 +37,6 @@ public partial class BlueskyAgent
         return await BlueskyServer.ListNotifications(
             limit,
             cursor,
-            seenAt,
             reasons,
             service: Service,
             accessCredentials: Credentials,